Add files via upload
using shortcodes-[add_to_my_list_button]
Function for button to be a different color to indicate it has already been added to avoid adding twice (such as turning the button grey with white text), or perhaps removing the button once it has been added and instead text that just says "Added"? I prefer the second option.
Instructions for how to take the packaged code and deploy it to two other live sites that we have for customers.
I didn't ask this before, but if the Add to my List button can be centered in the column, that would be an added plus and make it look much better.
If the Add to my List button can be the same color as the datatable header, and with white text, that would be great.